Move EmojiToolbar and Picker into /components
